pyproject.toml
[project]
name = "hansel"
version = "0.1.0"
description = "Autonomous AI agent that reads your CV, finds matching jobs, and drafts personalized application emails — running locally with Ollama."
readme = "README.md"
authors = [
    { name = "Alexia", email = "alexiahj111@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "cachetools>=7.0.6",
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "langchain>=1.2.15",
    "langchain-core>=1.3.0",
    "langchain-ollama>=1.1.0",
    "numpy>=2.4.4",
    "pdfplumber>=0.11.9",
    "pydantic>=2.13.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "tenacity>=9.1.4",
    "uvicorn[standard]>=0.45.0",
]

[project.scripts]
hansel = "hansel:main"

[build-system]
requires = ["uv_build>=0.8.23,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "jupyterlab>=4.5.6",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "respx>=0.23.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]
